Crate aoc_2016_day_09

Advent of Code 2016: Day 9: Explosives in Cyberspace

https://adventofcode.com/2016/day/9

Input

A stream of capital letters or repetition markers in the form (2x3), where 2 is the length and 3 is the number of repetitions.

Part one

Determine the uncompressed length, assuming the repetitions apply to the compressed stream directly, without recursively expanding nested repetitions inside the repeated portion.

Part two

Same as part one but recursively expanded nested repeated sections.

Algorithm

For part one we simply count letters directly, or count len × num for repeated sections, skipping the appropiate amount of data from the stream.

For part two, we keep a stack of repeated sections, along with the number of repetitions at the current level, and the position in the stream where each repeated section ends.

Modules

input

Data structures to model the puzzle input

logic

Functions

puzzle_metadata